# Add controllerUID index to improve ReplicaSetController performance